+++ d4np/d4np_c.h
@@ -26,6 +26,12 @@
 #include "d4np/mem/arena.h"
 #include "d4np/mem/pool.h"
 
+/* Concurrency & synchronization */
+#include "d4np/concurrency/atomic_queue.h"
+#include "d4np/concurrency/mutex.h"
+#include "d4np/concurrency/semaphore.h"
+#include "d4np/concurrency/thread_pool.h"
+
 /* Generic data structures */
 #include "d4np/ds/hashmap.h"
 #include "d4np/ds/linked_list.h"