+++ tests/tests.h
@@ -19,5 +19,9 @@ void suite_string_builder(void);
 void suite_hashmap(void);
 void suite_linked_list(void);
 void suite_ring_buffer(void);
+void suite_mutex(void);
+void suite_semaphore(void);
+void suite_atomic_queue(void);
+void suite_thread_pool(void);
 
 #endif /* D4NP_TESTS_H */